+++ ci/build.sh
@@ -8,7 +8,7 @@ function build_osx()
     echo "Building osx..."
     NUM_OF_CORES=`getconf _NPROCESSORS_ONLN`
 
-    cmake -GXcode -Bbuild -DYASIO_SSL_BACKEND=1 -DYASIO_HAVE_CARES=ON
+    cmake -GXcode -Bbuild -DYASIO_SSL_BACKEND=1 -DYASIO_USE_CARES=ON
     cmake --build build --config Release
     
     echo "run test tcptest on osx ..."
@@ -27,10 +27,10 @@ function build_ios()
 {  
     echo "Building iOS..."
 
-    # cmake .. -GXcode -Bbuild_ios -DCMAKE_SYSTEM_NAME=iOS -DCMAKE_OSX_SYSROOT=iphonesimulator -DYASIO_SSL_BACKEND=1 -DYASIO_HAVE_CARES=ON
+    # cmake .. -GXcode -Bbuild_ios -DCMAKE_SYSTEM_NAME=iOS -DCMAKE_OSX_SYSROOT=iphonesimulator -DYASIO_SSL_BACKEND=1 -DYASIO_USE_CARES=ON
     # cmake official cross-compiling doesn't works on latest macOS 10.15 + xcode11
     # https://cmake.org/cmake/help/v3.17/manual/cmake-toolchains.7.html?highlight=iphoneos#cross-compiling-for-ios-tvos-or-watchos
-    cmake -GXcode -Bbuild "-DCMAKE_TOOLCHAIN_FILE=$YASIO_ROOT/cmake/ios.mini.cmake" "-DCMAKE_OSX_ARCHITECTURES=arm64" -DYASIO_SSL_BACKEND=1 -DYASIO_HAVE_CARES=ON
+    cmake -GXcode -Bbuild "-DCMAKE_TOOLCHAIN_FILE=$YASIO_ROOT/cmake/ios.mini.cmake" "-DCMAKE_OSX_ARCHITECTURES=arm64" -DYASIO_SSL_BACKEND=1 -DYASIO_USE_CARES=ON
     cmake --build build --config Release
     
     exit 0
@@ -39,7 +39,7 @@ function build_ios()
 function build_linux()
 {
     echo "Building linux..."
-    cmake -G "Unix Makefiles" -Bbuild -DCMAKE_BUILD_TYPE=Release -DYASIO_SSL_BACKEND=2 -DYASIO_HAVE_KCP=ON -DYASIO_HAVE_CARES=ON -DYASIO_ENABLE_ARES_PROFILER=ON -DYAISO_BUILD_NI=YES -DCXX_STD=17 -DYASIO_BUILD_WITH_LUA=ON -DBUILD_SHARED_LIBS=ON
+    cmake -G "Unix Makefiles" -Bbuild -DCMAKE_BUILD_TYPE=Release -DYASIO_SSL_BACKEND=2 -DYASIO_USE_CARES=ON -DYASIO_ENABLE_ARES_PROFILER=ON -DYAISO_BUILD_NI=YES -DCXX_STD=17 -DYASIO_BUILD_WITH_LUA=ON -DBUILD_SHARED_LIBS=ON
 
     cmake --build build -- -j `nproc`
     
@@ -62,7 +62,7 @@ function build_android()
 {
     echo "Building android..."
     NINJA_PATH=~/ninja
-    cmake -G "Ninja" -B build -DANDROID_STL=c++_shared -DCMAKE_MAKE_PROGRAM=$NINJA_PATH -DCMAKE_TOOLCHAIN_FILE=~/android-ndk-$NDK_VER/build/cmake/android.toolchain.cmake -DCMAKE_ANDROID_NDK_TOOLCHAIN_VERSION=clang -DANDROID_ABI=$BUILD_ARCH -DCMAKE_BUILD_TYPE=Release -DYASIO_SSL_BACKEND=1 -DYASIO_HAVE_CARES=ON
+    cmake -G "Ninja" -B build -DANDROID_STL=c++_shared -DCMAKE_MAKE_PROGRAM=$NINJA_PATH -DCMAKE_TOOLCHAIN_FILE=~/android-ndk-$NDK_VER/build/cmake/android.toolchain.cmake -DCMAKE_ANDROID_NDK_TOOLCHAIN_VERSION=clang -DANDROID_ABI=$BUILD_ARCH -DCMAKE_BUILD_TYPE=Release -DYASIO_SSL_BACKEND=1 -DYASIO_USE_CARES=ON
     cmake --build build --target yasio
     cmake --build build --target yasio_http